requirements.txt
rpi.gpio
mingus